requirements.txt

langchain_openai
langchain_core
python-dotenv
huggingface_endpoint
streamlit
lagchain_community
langserve